requirements.txt
flask
tweepy==3.2.0
simplejson
numpy
python-firebase
gunicorn